requirements.txt
streamlit>=1.28.0
openai>=1.3.0
Pillow>=10.0.0
requests>=2.31.0
numpy>=1.24.0
reportlab>=4.0.4
python-dotenv>=1.0.0
google-generativeai>=0.3.0
streamlit-drawable-canvas>=0.9.3